refactor: split placesearch
