+++ app/PlaceSearch/_components/SearchResults/SearchHistory.tsx
@@ -1,0 +1,57 @@
+import Image from 'next/image'
+import SearchResultsContainer from './SearchResultsContainer'
+import { useLocalStorage } from 'usehooks-ts'
+
+export default ({
+	sideSheet,
+	onDestinationChange,
+	searchHistory,
+	setSearchHistory,
+}) => {
+	return (
+		<SearchResultsContainer $sideSheet={sideSheet}>
+			<div
+				css={`
+					display: flex;
+					justify-content: center;
+					align-items: center;
+					img {
+						width: 0.9rem;
+						height: auto;
+						vertical-align: sub;
+					}
+				`}
+			>
+				<small>Historique</small>
+				<button
+					onClick={() => setSearchHistory([])}
+					title="Effacer l'historique"
+				>
+					<Image src="/trash.svg" width="10" height="10" alt="Icône poubelle" />
+				</button>
+			</div>
+			<ul
+				css={`
+					li:not(:last-of-type) {
+						border-bottom: 1px solid var(--lightestColor);
+					}
+					li {
+						padding: 0.4rem 0;
+						list-style-type: none;
+						font-size: 90%;
+						line-height: 130%;
+					}
+					li:hover {
+						background: var(--lightestColor);
+					}
+				`}
+			>
+				{searchHistory.map((entry) => (
+					<li key={entry}>
+						<button onClick={() => onDestinationChange(entry)}>{entry}</button>
+					</li>
+				))}
+			</ul>
+		</SearchResultsContainer>
+	)
+}
