+++ app/PlaceSearch/_components/SearchResults/SearchNoResults.tsx
@@ -1,0 +1,9 @@
+export default ({ value }) => (
+	<div
+		css={`
+			text-align: center;
+		`}
+	>
+		<i>Aucun résultat pour la recherche “{value}”</i>.
+	</div>
+)
