+++ app/PlaceSearch/index.tsx
@@ -1,0 +1,236 @@
+import GeoInputOptions from '@/components/GeoInputOptions'
+import { InputStyle } from '@/components/InputStyle'
+import { getArrayIndex, replaceArrayIndex } from '@/components/utils/utils'
+import { buildAddress } from '@/components/Address'
+import fetchPhoton from '@/components/fetchPhoton'
+import { isIOS } from '@react-aria/utils'
+import Image from 'next/image'
+import Link from 'next/link'
+import { useCallback, useEffect, useRef, useState } from 'react'
+import styled from 'styled-components'
+import { useLocalStorage } from 'usehooks-ts'
+import { buildAllezPart, setAllezPart } from '../SetDestination'
+import { encodePlace } from '../utils'
+import { Loader } from '@/components/loader'
+import LogoCarteApp from './_components/LogoCarteApp'
+import SearchBar from './_components/SearchBar'
+import SearchHistory from './_components/SearchResults/SearchHistory'
+import SearchLoader from './_components/SearchResults/SearchLoader'
+import SearchNoResults from './_components/SearchResults/SearchNoResults'
+import SearchHereButton from './_components/SearchResults/SearchHereButton'
+import SearchResultsContainer from './_components/SearchResults/SearchResultsContainer'
+import { Geolocate } from './_components/Geolocate'
+import { FromHereLink } from './_components/FromHereLink'
+
+/* I'm  not sure of the interest to attache `results` to each state step.
+ * It could be cached across the app. No need to re-query photon for identical
+ * queries too.*/
+export default function PlaceSearch({
+	state,
+	setState,
+	sideSheet,
+	setSnap,
+	zoom,
+	setSearchParams,
+	searchParams,
+	autoFocus = false,
+	stepIndex,
+	geolocation,
+	placeholder,
+}) {
+	if (stepIndex == null) throw new Error('Step index necessary')
+	const [localSearch, setLocalSearch] = useState(true)
+	const [searchHistory, setSearchHistory] = useLocalStorage('searchHistory', [])
+
+	const urlSearchQuery = searchParams.q
+
+	const step = getArrayIndex(state, stepIndex) || {
+		results: [],
+		inputValue: '',
+	}
+	const value = step.inputValue
+
+	// hence this, but does it work with autofocus from triggered indirectly by
+	// the Steps component ?
+	const [isMyInputFocused, instantaneousSetIsMyInputFocused] = useState(false)
+
+	//
+	useEffect(() => {
+		// the onClick below doesn't work on iOS, and this technique produces worse
+		// results on Android Firefox and especially chrome
+		if (!isIOS()) return
+		if (!isMyInputFocused) return
+		setSnap(0, 'PlaceSearch')
+	}, [isMyInputFocused, setSnap])
+
+	const setIsMyInputFocused = (value) => {
+		// click on suggestion would close the suggestion before it works haha dunno
+		// why
+		setTimeout(() => instantaneousSetIsMyInputFocused(value), 300)
+	}
+
+	const onInputChange =
+		(stepIndex = -1, localSearch = false) =>
+		(v) => {
+			const oldStateEntry = state[stepIndex]
+			const stateEntry = {
+				...(oldStateEntry?.results?.length && v == null
+					? { results: null }
+					: {}),
+				...(v === '' ? {} : { inputValue: v }),
+			}
+			const safeStateEntry =
+				Object.keys(stateEntry).length > 0 ? stateEntry : null
+
+			console.log('will call replaceArrayIndex', state, stepIndex)
+			const newState = replaceArrayIndex(
+				state,
+				stepIndex,
+				safeStateEntry
+				//validated: false, // TODO was important or not ? could be stored in each state array entries and calculated ?
+			)
+			setState(newState)
+			if (v?.length > 2) {
+				const hash = window.location.hash,
+					local = hash && hash.split('/').slice(1, 3)
+
+				fetchPhoton(v, setState, stepIndex, localSearch && local, zoom)
+			}
+		}
+
+	const onDestinationChange = onInputChange(stepIndex, localSearch)
+
+	useEffect(() => {
+		if (!urlSearchQuery || value) return
+
+		onDestinationChange(urlSearchQuery)
+	}, [urlSearchQuery, onDestinationChange, value])
+
+	return (
+		<div>
+			<div
+				css={`
+					display: flex;
+					justify-content: center;
+					align-items: center;
+					margin-bottom: 0.4rem;
+
+					> button {
+						margin: 0;
+						padding: 0;
+						margin-right: 0.4rem;
+						> img {
+							width: 2rem;
+							height: auto;
+							vertical-align: middle;
+						}
+					}
+					margin-top: 0.2rem;
+					${sideSheet && `margin: .4rem 0`}
+				`}
+			>
+				<LogoCarteApp />
+				<SearchBar
+					state={state}
+					value={value}
+					onDestinationChange={onDestinationChange}
+					setIsMyInputFocused={setIsMyInputFocused}
+					setSnap={setSnap}
+					placeholder={placeholder}
+					autofocus={autoFocus}
+				/>
+			</div>
+			<div>
+				{state.length > 1 &&
+					stepIndex === 0 && // TODO for VIA geolocal, less prioritary
+					(geolocation ? (
+						<FromHereLink
+							geolocation={geolocation}
+							searchParams={searchParams}
+						/>
+					) : (
+						<Geolocate />
+					))}
+			</div>
+			{step.inputValue === '' &&
+				isMyInputFocused &&
+				searchHistory.length > 0 && (
+					<SearchHistory
+						onDestinationChange={onDestinationChange}
+						searchHistory={searchHistory}
+						setSearchHistory={setSearchHistory}
+						sideSheet={sideSheet}
+					/>
+				)}
+			{step.inputValue !== '' &&
+				(!step.choice || step.choice.inputValue !== step.inputValue) && (
+					<div>
+						{step.results ? (
+							<SearchResultsContainer $sideSheet={sideSheet}>
+								{step.results.length > 0 ? (
+									<GeoInputOptions
+										{...{
+											whichInput: 'vers', // legacy
+											data: step,
+											updateState: (newData) => {
+												setSnap(1, 'PlaceSearch')
+												// this is questionable; see first comment in this file
+												const newState = replaceArrayIndex(
+													state,
+													stepIndex,
+													newData
+												)
+												setState(newState)
+												setSearchHistory([
+													value,
+													...searchHistory.filter((entry) => entry !== value),
+												])
+
+												console.log('ici', newData)
+												const {
+													osmId,
+													featureType,
+													longitude,
+													latitude,
+													name,
+												} = newData.choice
+
+												const address = buildAddress(newData.choice, true)
+												const isOsmFeature = osmId && featureType
+												setSearchParams({
+													allez: setAllezPart(
+														stepIndex,
+														newState,
+														buildAllezPart(
+															name || address,
+															isOsmFeature
+																? encodePlace(featureType, osmId)
+																: '',
+															longitude,
+															latitude
+														)
+													),
+													q: undefined,
+												})
+											},
+										}}
+									/>
+								) : (
+									<SearchNoResults value={value} />
+								)}
+								<SearchHereButton
+									setLocalSearch={setLocalSearch}
+									localSearch={localSearch}
+									onInputChange={onInputChange}
+									state={state}
+									stepIndex={stepIndex}
+								/>
+							</SearchResultsContainer>
+						) : (
+							<SearchLoader />
+						)}
+					</div>
+				)}
+		</div>
+	)
+}
